#13: Split sort utils per resource into their own specific packages.
