+++ src/main/java/com/commercetools/sync/categories/utils/CategoryUpdateActionUtils.java
@@ -25,8 +25,8 @@ import javax.annotation.Nonnull;
 import java.util.List;
 import java.util.Optional;
 
+import static com.commercetools.sync.categories.utils.UpdateActionsSortUtils.sortAssetActions;
 import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
-import static com.commercetools.sync.internals.utils.UpdateActionsSortUtils.sortCategoryAssetActions;
 import static java.lang.String.format;
 import static java.util.Collections.emptyList;
