+++ src/main/java/com/commercetools/sync/products/utils/ProductVariantUpdateActionUtils.java
@@ -39,13 +39,13 @@ import static com.commercetools.sync.commons.utils.CollectionUtils.emptyIfNull;
 import static com.commercetools.sync.commons.utils.CollectionUtils.filterCollection;
 import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
 import static com.commercetools.sync.internals.utils.UnorderedCollectionSyncUtils.buildRemoveUpdateActions;
-import static com.commercetools.sync.internals.utils.UpdateActionsSortUtils.sortAttributeActions;
-import static com.commercetools.sync.internals.utils.UpdateActionsSortUtils.sortImageActions;
-import static com.commercetools.sync.internals.utils.UpdateActionsSortUtils.sortPriceActions;
-import static com.commercetools.sync.internals.utils.UpdateActionsSortUtils.sortProductVariantAssetActions;
 import static com.commercetools.sync.products.utils.ProductVariantAttributeUpdateActionUtils.ATTRIBUTE_NOT_IN_ATTRIBUTE_METADATA;
 import static com.commercetools.sync.products.utils.ProductVariantAttributeUpdateActionUtils.buildProductVariantAttributeUpdateAction;
 import static com.commercetools.sync.products.utils.ProductVariantPriceUpdateActionUtils.buildActions;
+import static com.commercetools.sync.products.utils.UpdateActionsSortUtils.sortAssetActions;
+import static com.commercetools.sync.products.utils.UpdateActionsSortUtils.sortAttributeActions;
+import static com.commercetools.sync.products.utils.UpdateActionsSortUtils.sortImageActions;
+import static com.commercetools.sync.products.utils.UpdateActionsSortUtils.sortPriceActions;
 import static java.lang.String.format;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
